Reorganise ledger to be more self contained
